+++ mlir/include/mlir/Dialect/MQTDyn/IR/MQTDynDialect.h
@@ -8,50 +8,38 @@
  * Licensed under the MIT License
  */
 
-#include "mlir/Dialect/MQTOpt/IR/MQTOptDialect.h" // IWYU pragma: associated
+#pragma once
 
-// The following headers are needed for some template instantiations.
-// IWYU pragma: begin_keep
-#include <llvm/ADT/TypeSwitch.h>
-#include <mlir/IR/Builders.h>
-#include <mlir/IR/DialectImplementation.h>
-// IWYU pragma: end_keep
+#include <mlir/Bytecode/BytecodeOpInterface.h>
+#include <mlir/IR/BuiltinAttributes.h>
+#include <mlir/IR/OpDefinition.h>
+#include <mlir/Interfaces/SideEffectInterfaces.h>
+
+#define DIALECT_NAME_MQTDYN "mqtdyn"
 
 //===----------------------------------------------------------------------===//
 // Dialect
 //===----------------------------------------------------------------------===//
 
-#include "mlir/Dialect/MQTOpt/IR/MQTOptOpsDialect.cpp.inc"
-
-void mqt::ir::opt::MQTOptDialect::initialize() {
-  // NOLINTNEXTLINE(clang-analyzer-core.StackAddressEscape)
-  addTypes<
-#define GET_TYPEDEF_LIST
-#include "mlir/Dialect/MQTOpt/IR/MQTOptOpsTypes.cpp.inc"
-      >();
-
-  addOperations<
-#define GET_OP_LIST
-#include "mlir/Dialect/MQTOpt/IR/MQTOptOps.cpp.inc"
-      >();
-}
+#include "mlir/Dialect/MQTDyn/IR/MQTDynOpsDialect.h.inc" // IWYU pragma: export
 
 //===----------------------------------------------------------------------===//
 // Types
 //===----------------------------------------------------------------------===//
 
 #define GET_TYPEDEF_CLASSES
-#include "mlir/Dialect/MQTOpt/IR/MQTOptOpsTypes.cpp.inc"
+#include "mlir/Dialect/MQTDyn/IR/MQTDynOpsTypes.h.inc" // IWYU pragma: export
 
 //===----------------------------------------------------------------------===//
 // Interfaces
 //===----------------------------------------------------------------------===//
 
-#include "mlir/Dialect/MQTOpt/IR/MQTOptInterfaces.cpp.inc"
+#include "mlir/Dialect/Common/IR/CommonTraits.h"         // IWYU pragma: export
+#include "mlir/Dialect/MQTDyn/IR/MQTDynInterfaces.h.inc" // IWYU pragma: export
 
 //===----------------------------------------------------------------------===//
 // Operations
 //===----------------------------------------------------------------------===//
 
 #define GET_OP_CLASSES
-#include "mlir/Dialect/MQTOpt/IR/MQTOptOps.cpp.inc"
+#include "mlir/Dialect/MQTDyn/IR/MQTDynOps.h.inc" // IWYU pragma: export